[Gradle Sync] Prevent crashes on logging with a disposed project
This change refactors the code that logs AndroidStudioEvents to the
UsageTracker and also ensures that we do not attempt to log an event
when the project has been disposed as this causes an exception to be
thrown when attempting to work out the sync type.

This change also adds comments to GradleSyncListener instructing
implementing classes that they need to handle the case where the project
is disposed.

Change-Id: I19165a955c58815834811cec7218df1351636951
Fixes: 117818398
Test: N/A