Fixing a bug where stop profiling would not stop the profiled session.
The reason for this bug is because the event group id (session id) is used as a
caching value. However the unified pipeline was not setting the group id on the
group before returning the results. This resulted in the SessionsManager missing
its cached value and re-creating all entries for the session. Tests were added to
both SessionsManagerTest and UnifiedEventsTableTest. The sessions manager was
missing a test to validate that when stop session is called the profiled session
is no longer alive. This test is needed because SessionManager maintains internal
states (selected session[cached], active session[polled], profiled session[cached]).
We test the first 2 in other test however we missed the profiled session bit.
The test for UnifiedEventsTable was updated to check that we set the event group id.
Both test fail without the UnifiedEventsTable change and both test pass with the
change.

Test: UnifiedEventsTableTest
Bug: 119267706
Change-Id: Ia1cd6ccfd50652b3fffeda7e5db5b625480ff3d1