Move ownership of CPU ABI arch to profilers-android
CPU ABI architecture is used by IntellijCodeNavigator for native navigation
and by IntellijProfilerServices to obtain the native symbols directories. In
the previous design, we were setting the ABI from CPU profiler, remembering
it in CodeNavigator (both in profilers module) and using it in the classes
mentioned above (both in profilers-android module).

By making the CPU ABI arch a Supplier and setting it in
AndroidProfilerToolWindow we can isolate the responsibility of handling it
to the profilers-android module, and we don't need to care about updating
the value when the architecture change, because the supplier will check the
selected process architecture everytime we need to use it.

Bug: 118865682
Test: Manual. go/native-symbolizer-test covers the symbolization scenarios.
      For code navigation, capture a C++ trace and double-click the function
      you want to navigate to. You should navigate to the parent function,
      specifically the line that calls the function you double-clicked.

Change-Id: Id70e2fdc5f2c4cedafe3153122a69cf99a108cb3