AssertionError: Qualified name is not allowed: @android:style/Theme.DeviceDefault
java.lang.AssertionError: Qualified name is not allowed: "@android:style/Theme.DeviceDefault"
while adding Blank wear activity to Watch Face project.
ThemeHelper was using a deprecated API to find the the theme style, and passing
the wrong Theme name (was removing "android:" from the theme name but not "@android:").
This CL refactors the code to use the new namespace api.

Bug: 116684080
Test: Added ThemeHelperTest
Change-Id: I8c6a3a33b372a98cbd9b8e5ee0c55c2ef2e9bd73