HPROF analyzer
When Android Studio detects it is running out of memory it will ask
the user to allow for creation and analysis of a heap dump.

The change includes a event-based HPROF parser and an analysis
component which walks the object graph in the heap dump and
produces a report. Such report can be sent to Google
for further analysis.

Report contains a number of sections sections:

1. Class histogram (top 50 by instance count and top 10 by memory use)
2. Picks 2-8 classes for which detailed report could help with leak
  investigation
3. For each instance of these classes, groups all shortest paths to GC
  roots and prints most common path, with metrics how many instances share
  each element of such path. Data is shown for each GC root that is on >30%
  of paths.
4. Summary of disposer tree. Instead of showing the whole tree, report contains
  counts of most common parent-child relationships in disposer.
5. Prints GC root paths to Disposable objects that has been disposed,
  but there is still strong reference to it.

Test: Added tests under com.android.tools.idea.diagnostics.hprof.*
Fixes: N/A
Change-Id: I9db2d3038951555b332ee3c7e3999c96a9ee9f62