Adds ExecutionTarget support for DeviceAndSnapshot.
In order to support synchronization of the Run/Debug/Stop buttons with
the new DeviceAndSnapshotComboBox device selection feature without hacking
around platform's execution code, we need to integrate with the built-in
ExecutionTarget framework.

This CL integrates with the ExecutionTarget framework by removing the
dependency on the default_target and providing our own
ExecutionTargetProvider with custom ExecutionTarges that map 1:1 to the
selected device in the combo box device selection UI/feature.

This CL also overrides Run/Debug actions' icons to also depend on the
selected device.

Change-Id: Ib7807ebff6f3299826d4201700986ff48ac18b35
Test: visually verified
Bug: 117829870