Disable asset creation for invalid res directories
com.android.tools.idea.actions.AndroidAssetStudioAction#isAvailable
doesn't check if the location of where the action is invoked references
valid res directories.  These asset creation options should not be
enabled when the res directories are invalid because the image asset
creation steps will autofill invalid folders to the new asset wizards.

This CL adds a step in #isAvailable that checks for a valid resource
directory in the module template paths of the invocation location of the
AndroidAssetStudioAction.

Change-Id: I38becadef35d302b0f8e086d5e3a635598d7aa46
Test-Strategy: ui.test.groups=DEFAULT
Bug: 62573412
Test: UI tests still pass.  Also manually went through several test
projects to make sure the UI is still smooth.