Add a Watermark concept to a PropertiesView.
A PropertiesView defines a properties panel. Add a concept of a
watermark which is something that is shown when the properties
panel is empty.

The watermark has:
- a message, saying why the panel is empty
- an action, saying how to populate it (optionally with a link
  to an action that enables the panel)
- a help message

This CL adds the watermark concept but the actions are not
implemented yet. In the CL see how it is used for the Nele layout
editor properties panel.

Change-Id: I81ef9275dd3b8d48b647f48c7eab8c796d424696
Fixes: 119349718
Test: Unit tests updated.