Increase zoom in jump for small values
When Zoom level is very low (10%), we try to make very small increments
on it. The problem is that on the design surface we have a check to
avoid doing very small increments. This causes the zoom in to stop
working in those cases.
This CL increases the jump size of zoom in since a bigger jump seems
adecuate.

Change-Id: I135cd1a3291482ae1af29f31f1d4ed8bafe47589
Fixes: 119211219
Test: Updated