Construct aar repositories in parallel
The time to construct the aar libraries grows linearly with the number
and size of the external libraries. Since this is invoked from the EDT
during various tasks, it leads to observable freezes the first time the
aar repositories are constructed.

This CL parallelizes the construction of the aars, which is trivial
since there are no dependencies between 2 libraries. In experiments, the
speedup seems linear (from 4s -> less than half a second) with the
number of threads.

Fixes: 119559205
Test: existing tests
Change-Id: Ib1ab246862338b26bcdce3e07b66f67a7a35003e