Improve the type mapping of android attribute types.
The types used for the Android attributes were not exact.
There is not enough information in attrs.xml to detect if a given
attribute is a color, color-list, or a drawable for example.

This CL improves the type mapping in TypeResolver by adding new
rules and a lot of exceptions.

The rules and exceptions were computed by an earlier attempt to
analyze the read calls performed in the android framework code
base. This CL takes the result of the analysis and use that
mapping as the expected test result see: AndroidAttributeFact.

Some of the results from the analysis did not have the exact
NelePropertyType and a number of hand edits were preformed to
get the expected mapping more accurate. The android
documentation has been used.

Bug: 119609627
Test: Unit test added.
Change-Id: I635ac6d83b731008e99baf676ec42e994c55a926