[Gradle Sync] Compare URLs instead of files when validating libraries
This change modified AndroidModuleDependenciesSetup#isLibraryValid to
only compare urls instead of files. This was causing us to always
recreate AARs since the res/ folder url that we get from Gradle is
sometimes empty which causes it to not register as a VirtualFile. As a
result we always treated these libraries as invalid.

Change-Id: I52d8c35b0dd72d688bee52fc04328ce9b9048c8e
Fixes: 119860928
Test: Unit tests added