Avoid checking cmake compiler caches
The CreateCppKotlinProjectTest creates a new project and checks if the
project files match the expected files. Recent work on CMake integration
caused this test to fail because the work introduced a new hidden
directory to the newly created project. This hidden directory does not
appear in the test project's directory when the test is ran locally,
however. This remains a mystery.

Bug: 119871858
Test: bazel test
//tools/adt/idea/android-uitests:CreateCppKotlinProjectTest
//tools/adt/idea/android-uitests:BuildCppKotlinTest

Change-Id: Idd7ec27295d6bbe0d397ee1f31e5f1c0d8069d9e