Make ViewAction an interface
Moved all the previous default behaviour to the AbstractViewAction class
and make it immutable.
Also, moved the re-ranking of actions to a ViewAction.withRank extension
that returns a new ViewAction with the given rank.

This is a step towards cleaning up the ViewAction system and integrating
it with the TargetActions.

Bug: 120009753
Test: Existing tests pass
Change-Id: I38c2597a4c44ec8b8a6cccf87f6897123db0a4d2