Extract asset rendering out of list renderer.
This CL introduce another inderection layer to render the assets' icons.
This enable to use a single ListCellRenderer for all assets and reuse
the rendering code for outside the list.

This also enable removing some of the responsabilies of the
ProjecResourceBrowserViewModel (name to be updated).

All consumer classes have been updated.

Test: Updated existing test and divided some of them to match the
refactoring
Bug: N/A
Change-Id: I34c545e8b93dc1a5f9f562007585946d5eddd109