Clean-up ActionManager API
This CL makes the ActionManager API a bit more consistent and makes
clear which methods are overridable. This should be split into an
interface and an abtract base implementation in a future CL.
Some functionality that is specific to one of the implementations has
been moved and the API simplified.

Bug: N/A
Test: Updated existing tests
Change-Id: I77414393a1c2ba1bc943661c2a961c3f7527c265