Add comment for why .cxx directory is ignored
As discussed in b/119871858, the .cxx directory can be safely ignored
since there are other integration tests that ensure that the builds with
and without the cache result in the same APK file.

The test that checks to ensure that the builds result in the same APK
file is CmakeAndroidGradleBuildExtensionsTest, under the tools/base
project.

Change-Id: Iada9d3fe8a204c5dfc67717d7fe4422d0e8d04f5
Fixes: 119871858
Test: CreateCppKotlinProjectTest, CmakeAndroidGradleBuildExtensionsTest