Remove naveditor dependency from NlModel
NlModel is part of common so it shouldn't depend on naveditor, as we're
planning to split the designer module into (at least) common, uibuilder and
navigation.

The dependency we're removing was caused because NlModel was checking if the
type of the model's file was NAV and using NavComponentHelper in case it
was. We can remove that by simply creating a generic method to register
NlComponents in common and call that instead.

Bug: 120120905
Test: updated affected tests
Test-Strategy: ui.test.groups=DEFAULT
Change-Id: If419139030d29b02d06ef9f7ad19e41d3753c4c8