Move all ActionTargets to ViewActions
Now all the old ActionTargets will be displayed as part of the popup
menu. This CL also removes the ActionTarget and ActionTargetGroup
classes since we should not be using them anymore.
In order to support this transition, there are some changes on the
ComponentAssistant handling since it will now trigger from the popup.

Test: Updated tests to work with the new targets
Change-Id: I481b4bc5faa54374493fd9bd443597a7b0b2cae9
Fixes: 120009753
Fixes: 119628608
Test-Strategy: ui.test.groups=PROJECT_SUPPORT,PROJECT_WIZARD,EDITING,DEFAULT