Register designer types instead of defining them statically
We were defining the types supported by designer tools in NlLayoutType,
which was part of the common module. That somewhat was restricting the
creation of new editors, because they would need to add the types they
support to NlLayoutType. Types need to define their corresponding resource
type and usually their toolbar action groups. It's natural that each editor
defines these on their own module/package, so we would be creating a
dependency from common to the editor module, which is not ideal. In fact,
NlLayoutType was causing the common module to depend on naveditor and
uibuilder.

In order to fix the issue, a registrar (DesignerTypeRegistrar) was created
in the common module, and we're using it to register types we want to
support. This way, new editors can depend on common and use this service to
register the types they support. Type definitions, including the toolbar
action groups and associate them with the corresponding resources, should
still be done within the editor module/subpackage.

This CL removes the last dependency from common to naveditor, allowing the
latter to be extracted to its own module.

Bug: 120409101
Test: Updated affected tests.
Change-Id: I3da1a3e3fa1557a5983683e802fa506831942a09
Test-Strategy: ui.test.groups=DEFAULT