Offer the option to disable SDK diff patching.
This CL exposes the option to disable SDK diff patching, which
might be beneficial for lower-end hardware. By default, diff
patching is on, meaning that Studio will attempt to optimize
SDK updates by download size, potentially incurring a higher
local installation cost (due to the need to apply patch as
opposed to just unzip & move as it would be in case of
a full installation). It has been reported that on lower-end
hardware applying the patch might take consirably longer than
a full download, plus it is not backgroundable.

Therefore offering the choice to disable diff patching is
a reasonable step to accommodate those cases.

Bug: b/67816403
Test: Added StudioSdkInstallerUtilTest to ensure correct factory is
created.
Change-Id: Ibc16caeb187fa749ba083b539174ddaf33defbc9