Introduce LegacyTarget interface
Adds the LegacyTarget interface to common module as a way to indicate
targets that can't be replaced by CommonDragTarget as of right now due to
design limitations. This way, CommonDragTarget#isSupported only needs to
check for this interface instead of worrying about every possible Targets,
which can even live in different modules.

Bug: N/A
Test: Current tests should pass.
Change-Id: I803637a9d9c8910b24733b7e13470ac44e36c008