Ignore empty selection events in DevicePicker
DevicePicker is the component responsible for tracking the user's device
selection and using that as the default selection the next time they go
to run their app. It registers itself as a ListSelectionListener of the
JBList containing all available devices and updates its persisted state
whenever the device selection changes. Any time the JBList's model is
set, however, this generates a valueChanged event with an empty
selection, clobbering the the cached selection that the user made. The
end result is that the device picker always defaults to the first
device in the list.

This has only become an issue since I4616f0759, which offloads AVD
loading to a background thread, displays all connected device options
in the dialog immediately, and then asynchronously updates the list of
selectable targets to include the AVDs once they're ready. Whereas
before we would choose a default device on the EDT before DevicePicker
was able to clobber the cached data in response to the model being set,
now we have to choose a default device again after the asynchronous
update, at which point the cached selection is gone.

This CL resolves the issue by having DevicePicker ignore valueChanged
events with an empty selection, as these could never correspond to an
actual choice made by the user.

Change-Id: I08758e02ef77695601e60091d93e61a33afa1320
Fixes: 120547166
Test: Manual