Fix the focus traversal in the property panel.
The new properties panel consist of tables mixed with editor
controls. The default focus traversal of the PTable handles
the table as a single tab stop i.e. the user <tab>s through
the editors in the basic section but the properties in the
declared table and the all attribute table seem to be skipped.

If an editor in the table has focus then the user <tab>s will
navigate through the editors in the table, but focus will
never leave the table.

Fix this by:
- forward a focus gain to the table to an editable row in the
  table.
- do not wrap the focus traversal inside the table.
- make sure to select the first focusable component in a table
  row when moving forward, and the last when moving backwards.

Bug: 120624148
Test: Unit tests added.

Change-Id: I874602d522aee54da177640e8f47165d8e09cd8e