Change Apply (Code) Changes to rely on real pkg name
This CL changes Apply (Code) Changes actions to check for
the real package name to determine if it should be enabled
or not (as opposed to the process name, which sometimes
differ from the package name if the user specifies the
"android:process" flag in the manifest).

This also fixes a long-standing bug with the Rerun/Debug
icons not being in the right state when we have an app
that uses the "android:process" flag. This also impacts
the old IR.

Bug: 119120361
Fixes: 71645350
Change-Id: Icf5aae9ab273912a6108403c34c2fd35944b047f
Test: N/A