Re-migrate DebugOnEmulatorTest into Bazel
The DebugOnEmulatorTest was taken out of Bazel because the test had a
dependency on the emulator. Unfortunately, the emulator is not available
on the pre-submit and post-submit bots, so building the test would fail,
even if the test never actually ran.

This change addresses this issue by making the dependency on the
emulator configurable. If the test runs with the option "--define
external_emulator=true", then DebugOnEmulatorTest's target will depend
on the emulator. When this option is absent, the target will not depend
on the external dependency. This allows the target to still build on
pre-submit and post-submit machines. Moreover, the test does not
actually run on any of the main pre-submit or post-submit machines, so
there is no need to actually run the test on such machines. The
configuration of such machines will not require the "--define..." option
in their invocations.

This reverts commit 0c0fb49dc9c5b2e2e8ba838495a58d9928ae7e8a, reversing
changes made to fa0a74369d29ea54539501c377d4d0fd0759e48c.

Bug: 112553978
Test: bazel test //tools/adt/idea/android-uitests:DebugOnEmulatorTest
Change-Id: I6a9ea8f05ce93433aa319f48aa0d0b2624dff71d