Remove NativeDebuggerGuiTestRule
Subclasses of GuiTestRule increase complexity and make it difficult to
reason about the scope of possible UI test behavior.

NativeDebuggerGuiTestRule customized GuiTestRule in three ways:
  1) it set a timeout of 5 minutes. This is already configurable in
GuiTestRule.
  2) it overrode getMasterProjectDirPath to point at the debugger
subdirectory. GuiTestRule's implementation calls
GuiTests.getTestProjectsRootDirPath, which contains logic to handle both
standard and release builds - the override did not call this, so would
be broken on release builds. This is a good example of the kinds of
interactions that make subclassing GuiTestRule a bad idea.
  3) it set the NDK path in an override of updateGradleVersions.

This change adds an option to GuiTestRule to set the NDK path, and
replaces usages of NativeDebuggerGuiTestRule with GuiTestRule, with a
timeout set to 5 minutes and the NDK path option set. It also prepends
"debugger/" to all of the importProject* calls in affected tests, to
address (2).

Bug: 120859192
Change-Id: I57918812adc03b269fee9275328ffb35d87905f3