Dispose AdbDeviceFileSystemService with project
AdbDeviceFileSystemService is created per project and expected to
be invoked on EDT thread. But it does not get disposed after project get
disposed. As a result, we can still access AdbDeviceFileSystemService
when project get closed and exception will be throw when we try to
access EDT thread i.e. when device state get changed.

With this patch AdbDeviceFileSystemService is registered as project's
child and they will be disposed together.

Bug: 120638225
Test: AdbDeviceFileSystemServiceTest.testServiceDisposed_afterParentDisposed
Change-Id: I0a08c6eda9c5385abb01c40d3dc70a38153b83c7