Add resource picker to drawable attributes.
We have a color picker for color types, but the color picker is
a surprising choice for a drawable that is already set to a
drawable value.
Change this such that both color and drawable pickers are available
for now: use the resource picker. Later implement an inline drawable
picker.

Change-Id: I45475edf0486cb8f19986cd93e7695f6e9ac8083
Fixes: 120655631
Test: Existing unit test + added a couple of extra tests.