Fix default value cache problem.
NeleDefaultPropertyProvider was caching the default value maps
created by layoutlib where the keys are the snapshot view instances
from layoutlib. Unfortunately the snapshots are recreated every
time we render, so we should not cache these maps.

In addition we also need to detect when the default values we show
in the UI are changed, and request the UI to update.

Bug: 121217307
Test: Added unit test for detecting an update.
Change-Id: I71d9687d4e124733c011b365a5d96e43b72d7ca6