Offer to use JAVA_HOME as JDK location
We are currently recommending to use the embedded jdk as the jdk
location when calling gradle from Android Studio, but by default Gradle
tries to use JAVA_HOME, this means that if the user runs Gradle task
using the CLI it is possible that 2 different daemons are running since
they do not share the same configuration (one using the embedded JDK
while the other uses the one at JAVA_HOME).

This CL adds an option to use JAVA_HOME as the JDK location, helping on
reducing the number of daemons. The changes here involve the first run
welcome wizard and the PSD to also have this option available on
subsequent Android Studio runs.

See these images to see the changes in UI:
 - PSD: https://screenshot.googleplex.com/LHXWUrnN8yF
 - First run (old wizard): https://screenshot.googleplex.com/fXhyRAPxCga

Only old version of the first run wizard includes the changes.

Bug: 115925397, 68374709
Test: None yet, want to get feedback on this first.
Change-Id: Ifcb9b6bf824a8744284d4b43c86f6ddb2fd9d0e0