Use the ResourceRepositoryManager for analyzing fonts.
Before this CL we have used the ResourceResolver to locate the
file of a downloadable font.
The ResourceResolver is expensive to create, and it is not
recreated after a font is added to the font resource folder.
The ResourceRepositoryManager has instant access to newly added
resources, so use that instead.

Change-Id: Ie8b63269f313dc792eef56b7a3bc43603219c1f2
Fixes: 122033860
Test: Existing tests shouls pass.