Refactor NelePropertiesProvider to load from styleables.
Properties can be found:
1) From XML using the extensions studio have added.
2) Going through all the styleable for a component and its
   super classes and the layout styleables for the parent layout.

So far we have done (1) and (2) for the inflated class which will
give us all attributes. But we are unable to determine which
component class an attribute from (1) were originally declared.
We need that information to e.g. create help links.

With (2) we can get this information but it will not give us all
the generated attributes we get from (1).

This CL implements both (1) and (2).
Most attributes are found twice but a local time test did not show
a significant decrease in performance. The time it took to find all
attributes was about 4 msecs (after the initial 1sec to load the
attribute definitions) with or without (2).

The extra information we need from (2) is not captured in this CL.
Here we are only enabling to capture the information in a follow-up
CL.

Bug: 121259587
Test: The functionality should be the same, so existing tests
      should pass.
Change-Id: Ib185145c3b7ed8bd344c5513f9f642d31e9797ec