Abandon closing modal dialogs when one is not showing
Many infinite modal dialog loops appear to be due to dialogs which are
not showing, yet still 'active'. When this happens, it also seems not to
affect subsequent tests (i.e., an IDE restart does not occur due to the
absence of the welcome frame).

This also removes the timeout on the modal dialog check loop, which is
(speculatively) no longer necessary.

Change-Id: Id584b0b95ae4a127cc4eb5c7ec3a62a6d407feb0