Refactor HelpSupport on a PropertyItem.
Instead of making a PropertyItem implementation implement
HelpSupport, make HelpSupport a field of PropertyItem.
This makes the feature more discoverable, and removes the
need to implement extra interfaces.

Also: change HelpSupport for help and secondaryHelp to
supply AnAction. This allows an event to pass crucial context
data to the help implementation. It also helps with isolating
the help implementation in a action class.

Minor: Isolate the help key registration in a HelpSupportBinding
class.

Bug: 121259587
Test: NA
Change-Id: I454e278a14878f324b89e545f28da21f4ff80784