Fix IdeFrameFixture.selectApp
Change Idde5a94 set StudioFlags.JVMTI_REFRESH to true, which caused
an additional ComboBox to appear in the action toolbar. The fixture simply
found the only ComboBox in the toolbar, which is now ambiguous. This
change makes the search more specific by looking for the ComboBox with
the correct tooltip.

ComboBoxActionFixture.findComboBoxByText was unused, and was replaced
with findComboBoxByTooltip.

Bug: 122276901
Change-Id: I4b7c18069a74acff7f5690f80b39abac98e81689