Guard against project disposal in DependencyManager.
In change: I1d1f1d we move the initialization onto a background
thread. The user may decide to close the project immediately
which would cause the crash from b/122373349 since the project
may not be valid when the message bus listener is setup.
Guard aganst that.

Also add a check in MigrateToAndroidxUtil.getProjectProperties
since this typically is called from a read transaction.
If the method is called from a background thread the project
may be disposed before the read transaction is started.

Change-Id: I4909a0992ee628bfb982559af305291150738374
Fixes: b/122373349
Test: unit test added.