Replace usage of Arrays$ArrayList with ArrayList
ConnectedDevicesWorkerDelegate#doInBackground() can return the output
of Arrays.asList(), which is an Arrays$ArrayList that does not support
the remove() operation.

This causes an exception to be thrown when elements are removed from the
collection in AsyncDevicesGetter#newVirtualDeviceIfItsConnected.

This CL simply wraps the Arrays#asList() call in a new ArrayList<>(),
allowing it to be mutable.

Bug: None.
Test: Manual.
Change-Id: I384ed0090e7a00951954e04a82967c750389d2d6