Avoid running potentially blocking ops on ForkJoinPool
The Layout Editor uses whenCompleteAsync that will run the task on
ForkJoinPool but the number of available threads is constant and might
be low. Some of the operations executed might be blocking and it
is not easy to control when they will block (since we use it to execute
listeners).
This change moves the execution of those async ops to the Application
thread pool.

Change-Id: Ib9a59912ea57b0e8e8f603957aca7baa5bbf96be
Fixes: 120866690
Test: Existing tests pass