Load project template in background thread and don't block UI thread.
Loading project templates performs disc access and parsing lots of XML files. In practice, this is not an issue for most of the time because there are only about 100 of small size XML files to parse. However depending on the Android Studio installation and project directory, this operation can be significantly slow such as in USB flash drive or slow network drive.

This CL moves the XML reading operation to the background thread and displays JBLoadingPanel while loading data.

Change-Id: Iec54b57ba64dca6b3ce817e8fac26afbd9a8e8af
Fixes: b/122085264
Test: CreateNewMobileProjectTest