Do the designer layout on resize even when no zoom is needed.
Resizing the designer in the layout editor normally keeps the
design image centered in the design area. However there are
certain conditions when this doesn't happen. When the design image
is small (e.g. because of the build toolwindow being visible),
resizing the design area does not keep the image centered.

Change-Id: I15bc1a75b2ac4cd2442ac5b65e731f4362174c58
Fixes: 120604370
Test: NA