Couldn't create more than one destination
Previously if you tried to create more than one destination in a graph,
the second and subsequent invocations didn't work.
Also fixing an infinite recursion that only seemed to manifest in dirty
test environments (that is, not in prod or on test robots).
Also fixing a problem where some render requests could be pending at the
end of a NavTestCase, resulting in them running on the normal
ThumbnailManager rather than through TestableThumbnailManager, which
would fail in bazel.
Also fixing the use of CompletableFuture in ThumbnailManager. Previously
if there were a failure during rendering (as in the aforementioned
case), the future used in RefinableImage would never be marked as
failed, and we would wait for it forever.

Test: this is it
Bug: 122896764
Change-Id: Ib7f03db55c2b3ca4ebea6c2af13f95dbfbaa0f1c