Fix IDE hangs when killing processes monitored on device
We kill processes being monitored on device by running the "am force-stop"
shell command through the IDevice#executeShellCommand method, which can take
a while to be executed.

Our crash reports (e.g. crash/1575067f3411197a) show that we were indeed
taking too long to kill the monitored processes, consequently freezing
Android Studio for several seconds because we were executing the shell
command in the UI thread.

This CL moves the execution of the "am force-stop" commands to the pooled
threads executor in order to fix the issue.

Bug: 122376044
Bug: 122820269
Test: N/A
Change-Id: I25d2020faaf8beb82870484fbffbfd69ba364448