Reduce IDE hang time when checking for multiple profiles
When checking if instant run is supported, we were indirectly checking the
device for multiple user profiles (e.g. work profile) and that's done by
invoking the "pm list users" shell command, which can take a while to run.
All of that is done in the UI thread, causing freezes that can take longer
than one minute (e.g. crash/1d2af97b92e4d73f) and lots of force quits.

MultiUserUtils#hasMultipleUsers receives a timeout and it's clear the
intention of the author was to timeout if the shell command didn't respond
on time but that would only occur if the shell command was run in another
thread, which is not the case. Therefore, adding a latch#await with a
timeout only after the execution of the shell command is not fixing the
issue because the freeze happens before that gets to be called.

In this CL, we alleviate the issue by passing the timeout directly to the
executeShellCommand call, failing early in case the command doesn't return
within the expected time frame.

The proper solution, would be to invoke the shell command using a
ProgressIndicator (therefore, non-blocking) when a device is selected and
cache the result. MultiUserUtils#hasMultipleUsers should access the cached
result and return immediately. However, instant run is deprecated and this
code is scheduled to be removed from our codebase, so we don't need to
invest the time to address this now.

Bug: 122376044
Test: Updated MultiUserUtilsTest
Change-Id: I8572b097d2714e99cf61e23c1a9ea2b4668f3239