Prevents IRv2 hyperlink text from being printed
The error message was being used for both the notification
as well as being printed to the console. However, since the
notification has extra capabilities to create hyperlinks
for the user to click on (and only useful in a notification
UI context), this hyperlink text was also polluting the
console output window.

This message has now been split into two different messages,
one for the notification ballon, and one for the console.

Change-Id: Ice38a1686bef9583cf99d4be9d8941a9dc7fbba6
Fixes: 122047445
Test: N/A