Add ability to cache values in EnumSupport.
The items for style and textAppearance relies on a relative slow
method of getting the items. The UI will call repeatedly to get
the items from the model and the items do not change until a
change is detected in overall model.
Add a way to cache these values until something has changed.

Change-Id: Iabca62e2063f0c1f330abe17359d036f37c41d07
Fixes: 122533253
Test: NA