Fill default Jdk Location on JdkSetupStep
Current default option for JDK location is to use embedded JDK, but
JdkSetupStep on the first run wizard currently is not setting the text
field to the actual value, causing it to be empty and preventing the
user to continue.

This CL sets this field accodring to the default value. Also adds a
section in the summary to show what the user selected.

Bug: 122885709
Test: Visual inspection of changes in UI
Change-Id: I409c2cb26ff7d1fe77857da18f7eede7d46ea82e