Add "Build Output" window support for legacy data binding errors
A recent CL added "Build Output" window support for a newer
JSON-encoded error message format, but users of older Gradle
plugins would not get to benefit from that, and updating
a codebase's Gradle plugin is not always viable. Therefore,
this CL adds support for the legacy databinding error format
as well.

Bug: 111411525
Test: Added
Change-Id: I64421bb1ff505bd6d03659fe3ad5558a83a06758