Show a progress indicator while processes are being killed
We kill processes in a separate thread because we don't wanna block the UI
but we don't care much about the result. However, it's probably useful to
show a progress indicator while the app is not killed yet so the user is
aware there is an alive thread trying to do so.

Bug: 122376044
Test: b/122820269 is tracking tests around killing the processes in a
      non-blocking manner.

Change-Id: I4653fac8315df9aba03a6d01121146940a2b6cb9