Remove parenthesis in BUILD file causing iml-to-build to fail.
iml-to-build is currently unable to parse tuples or expressions with
parentheses around it. As a result having those in BUILD files will
cause iml-to-build to fail.

This CL removes the parenthesis to temporarily resolve this issue. Note
that this is only a workaround to unblock iml-to-build from crashing and
does not fix the underlying issue.

Bug: 123051383
Test: iml-to-build works now.
Change-Id: I431eeebac381ff289c203f997f883e9526058dd5