Fix ConcurrentModificationException from palette.
There was an intermittent unit test failure with
ConcurrentModificationException when running
DependencyManagerTest.testEnsureLibraryIsIncluded

It turns out that the problem is that the collections
in TestProjectSystem is not thread safe and this test
is making 2 consecutive updates. That problem is
fixed by waiting for the sync update to finish in the
test.

Change-Id: I40d70d53a87f788ef652a39340a789acd6050e5e
Fixes: 123027226
Test: Fixes intermittent unit test failure.